System dashboard: add Battery light options
@dragonGR:
That menu is fully hidden if you don't have config_multiColorBatteryLed
and config_intrusiveBatteryLed at all set to true in device tree.

If you have config_intrusiveBatteryLed set to true but not config_multiColorBatteryLed,
then you get the battery led menu but without the color options.

So, if you have a battery led but no multicolor, just set config_intrusiveBatteryLed to true to get the menu.
It'll show only two options, one to disable the led if DND is enabled,and one to disable the blinking red led if battery is low.